requirements.txt
streamlit
pandas
openpyxl
pinecone>=6.0.0
google-genai>=1.0.0
python-dotenv
matplotlib
seaborn
